Add files via upload
Adding anchoring vignettes to questionnaires provides a method of adjusting estimates of a group difference on a trait for differences in response style i.e. accounting for a bias in responding in one group compared to the other.  This study simulates item and vignette responses for 4 items for two groups of respondents, each of n=500.  The study compares the performance of different response style models under different vignette presentation designs and different response style scenarios.
The zip files contain Stata script for an example analysis and output (log) file and Stata script for a very extended simulation study that focusses on unbiased estimation of the group difference in the means of the trait.